clean up some utils

+++ utils/iris.py
@@ -2,34 +2,31 @@ from __future__ import annotations
 
 import copy
 import json
+import pathlib
 
 import rasterio
 
 import config
-from utils.type import JSONLike
+import utils.file
 
 
-class ConfigurationFile:
+class IRISConfigurationFile:
     def __init__(self) -> None:
-        self._base = _read_base()
+        with pathlib.Path(config.var("IRIS_BASE_CONFIG_FILENAME")).open() as f:
+            self._base = json.load(f)
 
     def create(self, obj_id: str) -> None:
         img_path = f"{config.env('TEMP_DIR')}{obj_id}.stack{'.tif'}"
         out_path = f"{config.env('TEMP_DIR')}{obj_id}.iris{'.json'}"
-
+        if utils.file.exists(out_path):
+            return
         f: rasterio.io.DatasetReader
+        g: rasterio.io.DatasetWriter
         with rasterio.open(img_path) as f:
             # NOTE: IRIS expects image shapes in reverse NumPy order.
-            shape = [f.width, f.height]
-
-        j = copy.deepcopy(self._base)
-        j["images"]["shape"] = shape
-        j["segmentation"]["mask_area"] = [0, 0, *shape]
-
+            img_shape = [f.width, f.height]
+        out = copy.deepcopy(self._base)
+        out["images"]["shape"] = img_shape
+        out["segmentation"]["mask_area"] = [0, 0, *img_shape]
         with open(out_path, "w") as g:
-            json.dump(j, g)
-
-
-def _read_base() -> JSONLike:
-    with open(config.var("IRIS_BASE_CONFIG_FILENAME")) as f:
-        return json.load(f)
+            json.dump(out, g)
